requirements.txt
bottle==0.12.8
pingo==0.2.0
RPi.GPIO==0.5.11
